# appx-services/user-service/user-api-client/user-client/gen-src/api/openapi.yaml
openapi: 3.0.1
info:
  description: User API
  title: User
  version: "1"
servers:
- description: Local server
  url: http://localhost:8080/api
paths:
  /user:
    post:
      operationId: saveUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
      summary: Save User
      tags:
      - User
      x-content-type: application/json
      x-accepts:
      - application/json
  /user/{id}:
    delete:
      operationId: deleteUserById
      parameters:
      - description: Unique Id of an User
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "201":
          description: OK
      summary: Delete User By Id
      tags:
      - User
      x-accepts:
      - application/json
    get:
      operationId: getUserById
      parameters:
      - description: Unique Id of an User
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
      summary: Get User By Id
      tags:
      - User
      x-accepts:
      - application/json
components:
  schemas:
    User:
      description: Data element for User
      example:
        firstName: firstName
        lastName: lastName
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        email: email
      properties:
        id:
          format: uuid
          type: string
        email:
          type: string
        firstName:
          minLength: 3
          type: string
        lastName:
          type: string
      required:
      - email
      type: object
    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      type: object